+++ DependencyInjection/Configuration.php
@@ -17,8 +17,9 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
 
 /**
  * @author Joris van de Sande <joris.van.de.sande@freshheads.com>
+ * @final
  */
-final class Configuration implements ConfigurationInterface
+class Configuration implements ConfigurationInterface
 {
     public function getConfigTreeBuilder(): TreeBuilder
     {