+++ DependencyInjection/Configuration.php
@@ -15,25 +15,16 @@ namespace FH\Bundle\WebpackBundle\DependencyInjection;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
-use function method_exists;
-
 /**
  * @author Joris van de Sande <joris.van.de.sande@freshheads.com>
+ * @final
  */
 class Configuration implements ConfigurationInterface
 {
-    private const ROOT_NAME = 'fh_webpack';
-    
     public function getConfigTreeBuilder(): TreeBuilder
     {
-        $treeBuilder = new TreeBuilder(self::ROOT_NAME);
-
-        if (method_exists($treeBuilder, 'getRootNode')) {
-            $rootNode = $treeBuilder->getRootNode();
-        } else {
-            // BC layer for symfony/config 4.1 and older
-            $rootNode = $treeBuilder->root(self::ROOT_NAME);
-        }
+        $treeBuilder = new TreeBuilder('fh_webpack');
+        $rootNode = $treeBuilder->getRootNode();
         
         $rootNode
             ->children()