+++ FHWebpackBundle.php
@@ -15,6 +15,7 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
  * @author Joris van de Sande <joris.van.de.sande@freshheads.com>
+ * @final
  */
 class FHWebpackBundle extends Bundle
 {